+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-## <small>0.0.1 (2020-04-06)</small>
+## <small>0.0.1 (2020-04-07)</small>
 
+* Added missing Changelog ([e406856](https://github.com/Desmoswal/mttcsp-corp/commit/e406856))
 * Initial Commit ([732e055](https://github.com/Desmoswal/mttcsp-corp/commit/732e055))
 * Quick fix on readme and changed xenial to bionic in travis config ([386050c](https://github.com/Desmoswal/mttcsp-corp/commit/386050c))
 * Quick fix on test case, added bootstrap ([915dc59](https://github.com/Desmoswal/mttcsp-corp/commit/915dc59))
 * Quick fixes, change from scss to css, testing i18n ([eec2823](https://github.com/Desmoswal/mttcsp-corp/commit/eec2823))
 * Removed Bootstrap, App ready to work with TrulyUI ([1a5c2c7](https://github.com/Desmoswal/mttcsp-corp/commit/1a5c2c7))
+* Small fixes, added header and sidebar ([cf1e985](https://github.com/Desmoswal/mttcsp-corp/commit/cf1e985))